project: Add a Makefile
This Makefile is designed to be fast, readable, and extensible. The
compiler/linker flags are set unconditionally based on the needs of the
project and the hardware environment. LTO is used to reduce the padding
between strings in .rodata.

A valid configuration is required for all goals that compile code.
"clean" and "format" targets are exempted because they work the same for
any board/platform.

Signed-off-by: Samuel Holland <samuel@sholland.org>