include: Add additional standalone headers
This includes non-standard freestanding headers, miscellaneous utility
macros, and headers used to expose functionality from the compiler.

The assembly macros allow link-time garbage collection of unused
assembly code and data by placing each function and data object in a
separate section. This mirrors the C compiler's -ffunction-sections and
-fdata-sections options. The function macros also provide CFI and object
size metadata. The idea for these macros was taken from the ARM Trusted
Firmware.

compiler.h brings many of the more useful attributes, builtins,
keywords, and types into the C namespace. Part of them are standard;
others are not. While some of the attributes are meaningful in the
generated code (e.g. __alias, __weak, and noinline), most are just hints
to the compiler to catch bugs and improve optimization. Some of the
attributes and builtins are inspired by their use in the Linux kernel
source.

Signed-off-by: Samuel Holland <samuel@sholland.org>