arch: Add OpenRISC 1000 architecture definitions
These constants are defined as part of the OpenRISC 1000 architecture.
They are the same for all platforms based on that architecture, and are
useful to any software written for the architecture.

exception.h is simplified from the names given in the reference manual.
spr.h is cleaned up from the version submitted to newlib.

Signed-off-by: Samuel Holland <samuel@sholland.org>