project: Add MIT license to LICENSE.md
This is the license used by musl, where many of the C standard headers
are copied from. The headers from musl do not require a copy of the
license or attribution, per this note in musl's COPYRIGHT file:

    In addition, permission is hereby granted for all public header
    files (include/* and arch/*/bits/*) and crt files intended to be
    linked into applications (crt/*, ldso/dlstart.c, and
    arch/*/crt_arch.h) to omit the copyright notice and permission
    notice otherwise required by the license, and to use these files
    without any requirement of attribution. These files include
    substantial contributions from:

    Bobby Bingham
    John Spencer
    Nicholas J. Kain
    Rich Felker
    Richard Pennington
    Stefan Kristiansson
    Szabolcs Nagy

    all of whom have explicitly granted such permission.

However, the ctype.h header in this project contains functions copied
from musl's source code, which is released under the standard MIT terms.

Signed-off-by: Samuel Holland <samuel@sholland.org>