common: Implement the SCPI protocol framework
Communication uses the message box hardware and a pair of shared memory
areas. The framework is designed to be easily extensible to various
kinds of commands, as well as handling multiple concurrent commands.

Initially implement only the SCPI_CMD_SCP_READY command, as ATF waits
for this command before continuing the boot process.

Signed-off-by: Samuel Holland <samuel@sholland.org>