Makefile: Generate a config header and make fragment
This allows code to use variables defined in the per-board config files,
such as the platform name.

Export the platform headers to tools, so they can take advantage of the
definitions they contain (memory/device addresses, etc.). Because this
allows tools to be platform-dependent, also export the board
configuration, including the platform name. That way, tools can report
which platform they were compiled for.

Update all existing boards to use the new format.

Signed-off-by: Samuel Holland <samuel@sholland.org>