drivers: clock: sunxi-ccu: Start bitmap at register base
Instead of adding the GATE/RESET offset to the register base, include it
in the bitmap index. This removes the platform/device-specific
dependency of the CCU driver that never should have been there.

It's not completely clear what to use in the BASE offset divisor.
(WORD_BIT / CHAR_BIT) would also work, and would be more akin to the
other bitmap code. However, the helper functions assume uint32_t.

Signed-off-by: Samuel Holland <samuel@sholland.org>