drivers: clock: sunxi-ccu: Rework driver data for flexibility
This expands the driver data to include a structure defining the
properties of each clock in the CCU. Instead of trying to pack
information about each clock into the device->clock field, each device
references the CCU's driver data by index. For the moment, only clock
gates and reset lines are included in the driver data, as these are the
only features supported by the driver.

Signed-off-by: Samuel Holland <samuel@sholland.org>