Makefile: Format all platform source files
$(sources) and $(includes) only include platform sources for the
currently-selected platform. This breaks for two reasons: we want to
make sure all files (not just the current platform) are correctly
formatted, and the CI build runs with no platform selected. Fix that by
wildcarding the platform directory. While here, change the file basename
patterns and add the tools directory, so no postprocessing is needed.

There's no good way to make this all fit in 79 characters, so allow 80.

Signed-off-by: Samuel Holland <samuel@sholland.org>