Makefile: Silence unhelpful initializer warning
gcc warns when some array or struct field initializers are missing. This
is not supposed to affect designated initializers, but when the order of
the fields in the device struct initializers are fixed, gcc warns about
them. Since the warning is just gcc reminding us that we may have
forgotten an initializer (we didn't), and it's not a code correctness or
conformance issue, silence the warning.

Signed-off-by: Samuel Holland <samuel@sholland.org>