debug: Add an option to print CPU SPRs
The OpenRISC 1000 architecture defines various Special Purpose
Registers, or SPRs, that describe the particular CPU implementation.
With this new option enabled, the firmware prints a subset of these SPRs
in a human-readable format during boot.

This information is used in the AR100 page on the linux-sunxi wiki, and
can be used to compare the hardware integrated into different SoCs.

Signed-off-by: Samuel Holland <samuel@sholland.org>