common: Implement a delay function using the wallclock
This is useful for hardware where we can't change registers too fast.
For example, both the CCU and R_TIMER devices require delays in between
two register updates. In these cases, the delay required is much too
short for the overhead of a timer interrupt to be worth using the timer
framework.

Signed-off-by: Samuel Holland <samuel@sholland.org>