Makefile: Fix inclusion of config.mk with multiple goals
If multiple goals are specified on the command line, and any of them
require a platform, be sure to include config.mk. Do this even if other
goals (that do not require configuration) are given. The outer condition
handles the case when no goals are given on the command line, and the
inner case handles when any goals are given that require configuration.

Signed-off-by: Samuel Holland <samuel@sholland.org>