common: scpi: Fix buffer struct alignment issues
Because they must match the SCPI ABI (the layout in memory), these
structs were marked as __packed to prevent GCC from inserting padding.
Unfortunately, that also causes GCC to consider them to have an
alignment of 1. Thus, when `struct scpi_mem` is put in `struct
scpi_buffer` along with two other bytes of data, there is no padding
added to the end of the structure. (There should be two bytes of padding
added, to match the correct alignment of 4). This then causes elements
of the `scpi_buffers` array to be misaligned, causing alignment faults.

Fix this by removing the __packed attribute, as the structures are
already laid out in a way that avoids padding between members. Then the
structures have the proper alignment, and padding is added at the end of
`struct scpi_buffer`.

Signed-off-by: Samuel Holland <samuel@sholland.org>