common: Add a complete system power control implementation
The system can be in one of several states; right now, these are off,
on, and suspended. Implement the functions for transitioning between
these states. Use PMIC functionality if it is available. If there is no
PMIC, or it is not functional, fall back to a fake off state and the
existing watchdog-based reset.

Try to save as much power as possible during shutdown by turning off all
non-critical regulators.

Partially based on a patch by Drew Walters.

Signed-off-by: Samuel Holland <samuel@sholland.org>