drivers: irqchip: sun4i-intc: Fix receiving NMIs (IRQ 0)
As an optimization, the current code tries to handle all incoming IRQs
during a single interrupt. This has worked before because the current
IRQ register reads as zero when no IRQs are pending. However, the NMI
pin connected to the PMIC is at interrupt zero. In that case, the
register will read zero even though there is a pending interrupt.

Solve the issue by only handling one IRQ per interrupt.