Makefile: Fix build with various toolchain options
GCC can be configured to generate position-independent (PIE) executables
by default. This is not necessary for a firmware whose address is
limited by hardware, makes the code larger and slower, and creates
relocations which must be processed at runtime. Pass -fno-pie and
-static to the compiler/linker to suppress this.

The technically correct option to pass to the linker is -no-pie, not
-static, but -no-pie is not recognized by older toolchains. Since we are
using a linker script, the other effects of -static don't apply, and it
ends up doing the same thing.

Pass --no-dynamic-linker to the linker to avoid an error about "not
enough room for program headers" triggered by newer versions of ld.

Signed-off-by: Samuel Holland <samuel@sholland.org>