project: Provide infrastructure to build kconfig
Kconfig comes with a yacc/lex (bison/flex) dependency, and the Makefile
needs new rules to use them. Then the Linux scripts/kconfig/Makefile is
adapted (and greatly simplified) to provide the standard *config goals.
Since we're very strict with warnings, but we really don't want to touch
this third-party code, turn off a few so it compiles cleanly.

A couple of other files are needed for the conf/nconf binaries to
actually run: "Kconfig" and "defconfig". For now the defconfig is empty.

Signed-off-by: Samuel Holland <samuel@sholland.org>