Makefile: Better way of deciding if config is needed
The configuration data is only needed if any compiling is requested.
clean/config/format goals do not require configuration, and may need to
run before configuration is available. The default goal ("all") does
involve running the compiler, so be sure to handle that case as well.

This construction makes it more clear why there are two conditions: the
default goal requires a functioning compiler.

Since this code is above the first goal, we must explicitly set the
default goal variable before using it.

Signed-off-by: Samuel Holland <samuel@sholland.org>