project: Incorporate a Kbuild-inspired build system
This build system implements a syntax for dependency specification
similar to that of Kbuild. Unlike Kbuild, it is implemented without any
recursive invocation of `make`, and without any use of the shell. Each
included Makefile is processed only once (in fact `make` enforces this).

In addition to providing a significant improvement in flexibility, this
system was benchmarked to be faster than the previous system that used
$(wildcard) on a list of subdirectories.

There are five top-level targets:
 - hostprogs-y: binaries used during the build (no implicit dependency)
 - lib-y: objects included in "lib.a", built for both host and target
 - obj-y: objects included in the main SCP firmware
 - test-y: binaries run as part of the unit test suite (`make check`)
 - tools-y: binaries made to run in Linux on the device (`make tools`)

Subdirectories are processed the same regardless of which target they
are added to. They are denoted by the subdirectory name followed by a
trailing slash, e.g. "foo-y += bar/".

See Documentation/kbuild/makefiles.txt in Linux for more information.

Signed-off-by: Samuel Holland <samuel@sholland.org>