Makefile: Move formatting code to a separate file
At the moment, it is intertwined with the functions used in the main
build system. Those functions are going away, and the formatting code
will be entirely separate. (Since the new build system does not always
compile every source file, we can't use the list of objects to
reconstruct the list of sources. Also, we want to exclude third-party
sources from formatting.)

Since the formatting code is independent and not really related to
_building_ the software, move it to a separate Makefile. Include this
new Makefile from the main one.

Signed-off-by: Samuel Holland <samuel@sholland.org>