common: Move the linker script to a more proper location
While the linker script is called a "script", it is more of a source
file. And like source files, it is run through the C pre-processor. It
becomes even more of a source file as it is used to define symbols
referenced from C and assembly code.

Move the file to make it clear that the scripts directly is only for
files used to transform source code, not inputs to the transformations.

While moving it, rename it to avoid scp/scpi tab-completion frustration.

Signed-off-by: Samuel Holland <samuel@sholland.org>