drivers: Move device definitions into each driver
Instead of defining all of the devices in one file, with a pile of
`#ifdef`s, put each device in the file implementing its driver. This
makes each device implicitly conditionally compiled with its driver,
without any preprocessor madness.

While headers are included unconditionally (even if the referenced
devices are compiled out), this is fine, as they are just extern
declarations.

This also has the side effect of putting all device structs into the
global namespace.

Signed-off-by: Samuel Holland <samuel@sholland.org>