lib: Clean up MMIO helpers
Rename them to be shorter, yet more readable. Fix the usage of volatile
(it applies to the variable type, not the type of usage, even though GCC
does what we want either way). Put mmio_get_32 in the firmware as well.

Changing poll(z) to inline does not affect generated code size.

Signed-off-by: Samuel Holland <samuel@sholland.org>