global: Make variable declarations more readable
In an attempt to improve clarity through alignment, the existing code
style rules aligned types as well as assignments. This proved to be
highly unreadable. While aligning the '=' in assignments is
overwhelmingly an improvement, there were practically zero cases where
aligning types in declarations was useful. In fact, several times, the
type alignment was so bad that I renamed variables or types to avoid it.

Reduce the threshold for type alignment to 1, to handle the case of
(u)int8_t next to its larger cousins, where the alignment still has
value.

In all other cases, declaration blocks generally follow "Reverse
Christmas Tree" style, with two specific rules:
  1) Structures come before all other declarations
  2) The error or return code variable comes last

Signed-off-by: Samuel Holland <samuel@sholland.org>