common: Optimize exception handling
These changes are intended to reduce the size of the exception handling
code, and to reduce the need for a separate stub for each handled
exception. The end goal is to remove the exception stubs entirely, and
perform all exception handling through a common entry point.

In order to save the exception number without using the stack, we need
an otherwise-unused register (or one we can safely clobber). r2 is a
callee-saved register. It is the last one in GCC's register allocation
order, as it normally holds the frame pointer. It is not used for that
purpose here, since the firmware is compiled with -fomit-frame-pointer.

Pass -ffixed-r2 to GCC to prevent it from ever allocating r2. This
affects exactly one place in the firmware where r2 is currently used.
Ironically, avoiding the allocation of r2 there actually makes that
function smaller.

Since GCC doesn't know that we saved all registers before calling
handle_exception, it generates code to spill callee-saved registers.
Take advantage of this, and only save call-clobbered registers in the
assembly.

Additionally, don't skip 128 bytes of stack, as GCC does not generate
code that uses the red zone. Unfortunately, there is no compiler flag to
make this ABI choice explicit.

Finally, only grab the EPCR SPR if it will be used in the panic message.
This allows removing the second argument to handle_exception without
losing much information.

Signed-off-by: Samuel Holland <samuel@sholland.org>