common: Fix the firmware to work above 64k
In order to save a few instructions, the hand-written assembly assumed
that the top 16 bits of memory addresses were always zero, and skipped
initializing them. This obviously only works if the entire firmware is
in the bottom 64k of address space.

Remove this arguably premature optimization to allow the firmware to be
loaded at any address within SRAM A2. Make sure to properly initialize
all pointers; and move the stack to immediately after .bss, instead of
right before the SCPI shared memory.

Signed-off-by: Samuel Holland <samuel@sholland.org>