common: Remove exception vectors from the firmware
Having exception vectors stored pre-calculated in the firmware image
forces the firmware to be loaded at the beginning of SRAM A2. To allow
flexibility in the load address (specifically, to allow moving the
firmware to the top half of SRAM A2), require the firmware loader to set
up the exception vectors prior to starting the firmware.

Assuming a single entry point, the calculation is straightforward:

  offset[n] = (entry address - vector address[n]) / 4

And since the opcode for l.j is (0x00000000 | offset), and the l.nop in
the branch delay slot is fixed in hardware, writing the exception
vectors simply requires writing the offset at each vector address. The
userspace firmware loader has been updated to perform this operation.

The PPC (previous program counter) SPR is used to recover the branch
source address. While this sort of use of the PPC SPR is discouraged by
the or1k architecture specification, it is the only known way to
calculate the exception number inside a shared handler. In order for the
l.mfspr to work, it must be the first instruction of the exception
handler, and thus the firmware. Therefore, the presence of a l.mfspr can
be used to check if a valid firmware was loaded.

It was experimentally determined that the contents of the PPC SPR
(unsurprisingly) followed the formula (0x100 * n + 8) for exceptions
that happened at runtime. However, at reset, the PPC SPR was observed to
hold the address of the l.mfspr instruction itself. This distinction is
used to separate firmware entry from exception handling.

Remove the definitions for the separate vector area, and merge them into
the SRAM A2 definitions.

Signed-off-by: Samuel Holland <samuel@sholland.org>