scpi: Update the SCPI client order
The new ABI allocates mailbox channels from high to low privilege,
resulting in the following map:

  0 =  S EL3 = ATF
  1 = NS EL2 = Linux
  2 =  S EL1 = TEE (future)

The final pair of mailbox channels is reserved for use by Linux with an
alternate protocol.

Shared memory regions follow the same pattern, but are allocated from high
to low memory addresses. This effectively allocates them in the opposite
order, but allows for future expansion to the third (TEE) SCPI client.

Since there is no TEE at the moment, no mailbox channel or shared memory
is reserved for it.

While swapping the SCPI memory regions, introduce a symbol provided by
the linker script to improve type safety.

Signed-off-by: Samuel Holland <samuel@sholland.org>