dm: Use an array for the device list, not a section
While using the linker script to colocate all of the device definitions
makes enabling/disabling and adding new devices a bit more convenient
(because they're only referenced in one place), it's non-obvious what
the magic `__device` attribute does.

More importantly, having a list of `struct device` *values* (as opposed
to pointers) prevents intrusive use of `struct device`. This is
inefficient, since it forces the structure to have several mostly-unused
members. It makes more sense for each driver to declare the structure
for only what it needs (device handles, configuration data, etc.).
Having a per-driver intrusive structure avoids the casts required by the
generic `drvdata`, and makes it easier to constify device definitions.

Only the top-level devices are included in the list: thost that provide
a SCPI service and those that are used internally by the firmware. All
others are initialized as dependencies of these devices; unused devices
(e.g. the I2C and RSB drivers on boards where those aren't used) can be
discarded with LTO.

Signed-off-by: Samuel Holland <samuel@sholland.org>