include: Fix bitfield macros
These were broken by 43e28b1a2d3a ("global: Use the GENMASK macro
instead of BITMASK"). To avoid confusion in the future, the uses of
GENMASK are simplified to use a single mask (instead of combining two).
This works because all instances of GENMASK in the file had a hardcoded
second parameter of zero.

Signed-off-by: Samuel Holland <samuel@sholland.org>