runtime: Disable use of missing/broken instructions
[commit 09652f82cf4fec6e3f35c51a98752b094bd95607 upstream]

Results from the ar100-fuzzer[1] show that the l.ext?? family of
instructions are not available on the AR100 (always returning zero). The
fuzzer also confirmed the suspicion that l.muli is broken. By providing
our own multiplication function, we can force the use of l.mul in all
cases. Unfortunately, GCC is not able to inline this function when
written using inline assembly in C, so it's actually smaller when using
hand-written assembly (GCC doesn't use the delay slot).

Signed-off-by: Samuel Holland <samuel@sholland.org>