common: Run sensor polling from the main loop
There's no real benefit to using a timer interrupt, when all it does is
add things to a queue. It provides no bound on latency, since if the
firmware is hung, it won't process the queue. So replace code that
schedules a timer interrupt every second with code that checks when the
wall clock has passed a one-second threshold. To maintain accuracy over
time, always add exactly one second to the threshold, regardless of how
long each loop iteration takes.

Signed-off-by: Samuel Holland <samuel@sholland.org>