scpi: Remove the payload for SCP-initiated messages
The only caller of this function (for SCP_READY) does not send a
payload. The payload would only be used for SCPI_CMD_ASYNC_SENSOR, which
is not supported. If it is supported in the future, the feature can be
added back. Removing thls for now reduces some noise in the SCPI rework.

Signed-off-by: Samuel Holland <samuel@sholland.org>