scpi: Rename the clients for clarity
The distinction between the two clients is not really the "secure"
status, but that certain commands are reserved for the PSCI
implementation in EL3. This rename makes it more obvious in the code why
the restrictions are in place. Plus it allows for a future Secure EL1
(TEE) client, which would conflict confusingly with the
SCPI_CLIENT_SECURE name.

Use an enumeration for better parser/type safety.

Signed-off-by: Samuel Holland <samuel@sholland.org>