scpi: Process requests asynchronously
Instead of performing the entire SCPI request/reply flow in the incoming
message handler, only mark the RX buffer as full at that point. Then,
in the polling loop, check for full buffers and try to flush them:
either by handling a request, or checking for acknowledgement of a
reply.

The main benefit of this change is that there are no longer any busy
loops or delays. Therefore, is is impossible for one client to block
requests from another by sending too many requests (we will attempt to
handle all outstanding requests before checking for more), or by failing
to acknowledge replies sent to it.

Since a TX buffer can no longer be filled asynchronously by IRQ handler,
the locking in scpi_create_message is not needed anymore.

Signed-off-by: Samuel Holland <samuel@sholland.org>