runtime: Fix result of __umodsi3 when modulus is 1
[commit 5311e2a1c71b88034b590ec477576c6532fe7938 upstream]

The __udivsi3 function has a special case to optimize the division
identity (division by 1). That code path did not initialize the
remainder in r12, so that function returned the dividend for the
remainder instead of the correct value, zero.

The fix is to clear r12 in the identity case.

This bug was found via code inspection; presently there are no callers
to __umodsi3 in the firmware, because all use of the modulus operator is
with power-of-two moduli (and that simplifies to a bitwise AND).

Signed-off-by: Samuel Holland <samuel@sholland.org>