dm: Rework IRQ descriptions to use less runtime modification
Make the `dev` member of `struct irq_handle` point to the controller,
like for every other class, instead of rewriting it to point to the
device with the IRQ handler. Pass handlers a pointer to the handle
itself, which they can use to find their device struct. This makes the
handle independent from the layout of `struct device`.

Signed-off-by: Samuel Holland <samuel@sholland.org>