project: Allow locally disabling formatting
The Kconfig struct/array initialization macros are a net benefit to
readability, but uncrustify cannot see the hidden commas, so it
misinterprets the address-of operator in later array items as the binary
AND operator, and thus adds extra unwanted spaces around it. The easy
solution is to turn off formatting for the affected initializers.

Signed-off-by: Samuel Holland <samuel@sholland.org>