drivers: clock: Make all clock device data intrusive
This increases type safety when defining the device data (though access
at runtime through container_of is not much better than casting
drvdata). It also makes way for removing the overhead of per-device data
on devices where it is not needed (by removing drvdata).

Unfortunately, there's no way around making the enclosing struct public
and using `.dev` everywhere. This should be acceptable, since one of the
goals of reorganizing the device structure is allowing (where necessary)
the device constants to be placed in a per-SoC file conditionally
compiled by Kconfig/Kbuild. In that case, since the driver and device
definitions would be in different files, the structure must be in a
header anyway.

Signed-off-by: Samuel Holland <samuel@sholland.org>