common: Reset the firmware after an exception
Since all exceptions are fatal, there is no need to save any registers
when handling an exception. The exception number is already available in
`start`, so we can send it as a parameter to `main` instead of branching
early. And by printing the error message in the normal startup path, the
exception path can be removed entirely.

As an added bonus, instead of spinning and waiting for the watchdog to
reset the system, the firmware can immediately attempt to recover and
continue running.

Note that while .bss is wiped on a reset, .data is not set back to its
original values. This doesn't appear to be problematic at the moment:
devices aren't reinitialized, but they don't need to be. This will be
fully resolved in the future as all data is moved into .rodata or .bss.

Signed-off-by: Samuel Holland <samuel@sholland.org>