drivers: pmic: Rename "wakeup" to "resume"
In this firmware's nomenclature, "wakeup" is the generic term for
initiating a power state transition from lower to higher power usage.
This PMIC hook is only used when resuming from suspend. Wakeup from an
off state is either handled internally to the PMIC (and the firmware
isn't running because the entire SoC is off), or it is implemented by
resetting the SoC in the same manner as a reboot.

Signed-off-by: Samuel Holland <samuel@sholland.org>