drivers: sun4i-intc: Add a default wakeup action
Except for a few possible future scenarios (such as only responding to a
specific keycode from the IR receiver), any incoming IRQ is going to
trigger a wakeup. Moreover, Linux needs to see the still-pending IRQ so
it can determine the correct wakeup reason. Leaving the IRQ pending is
fine because we stop polling as soon as a wakeup is triggered.

Signed-off-by: Samuel Holland <samuel@sholland.org>