global: Constify struct device
Everywhere that took a `struct device *` now should take a `const struct
device *`, and all of the definitions have been made const. The one part
of the structure that must be mutable is now accessed via a pointer.
While this wastes some space, the cost is not large (the mutable struct
has no padding), and there is a net gain of about 100 bytes.

Signed-off-by: Samuel Holland <samuel@sholland.org>