drivers: Remove IRQ handler registration code
Currently, the interrupt controller is only used to trigger a wakeup.
There are no users of this API. The only predicted use of an IRQ inside
the firmware, other than immediately triggering wakeup, is for the CIR
receiver. It could use an IRQ handler to limit wakeup to specific
keycodes. However, that could be coded directly, when a driver is
implemented.

The current code is problematic because it requires structures
containing both mutable and constant data, forcing the use of the .data
section. It is being removed as part of the constification effort.

Signed-off-by: Samuel Holland <samuel@sholland.org>