drivers: Make the counter a driver
This file was placed in the common directory back before Kconfig was
implemented. Now that a real configuration framework is in place, we can
put the various implementations in different files, and choose the right
one without linker magic or preprocessor ugliness.

For now, there is only one generic implementation, so enable it by
default.

Signed-off-by: Samuel Holland <samuel@sholland.org>