drivers: Use a driver for serial output
This was some of the first code in the firmware, in assembly out of
concern that the C environment may not be available. Now that the
startup code is well-debugged, replace the handmade assembly with
simple, readable C. Note that the console would not work pre-C anyway,
because it required an extra, unnecessary setup step. This change also
gets rid of a mutable object in .data.

Signed-off-by: Samuel Holland <samuel@sholland.org>