global: Use application-namespace identifiers for attributes
Identifiers starting with a double underscore are reserved for the
implementation (the compiler), not conforming applications. However, due
to conflicts (const, malloc, printf), the underscores cannot simply be
removed. They could be renamed from __foo to attribute_foo, but even
simpler is to use a macro to hide only the awful syntax, not the names
of the attributes -- that was the goal from the beginning.

Since __attribute__((unused)) is the most common attribute, and
__attribute__((weak)) is used on function definitions, they are provided
their own macros.

This change follows the rule of preferring attribute specification on
function prototypes, not function definitions, when possible.

Signed-off-by: Samuel Holland <samuel@sholland.org>