scpi: Poll each client channel individually
This simplifies the overall logic, since it doesn't go back and forth
between the msgbox driver and the SCPI service anymore. It also turns a
two-step polling loop into a one-step polling loop, which removes the
need for the rx_full flag (since we don't even look for new messages
until we have space to process one, all messages we receive are handled
immediately).

The msgbox polling handler is completely removed; if called, it would
drop messages from all channels. The one downside of this is that we
never flush the non-SCPI channels.

Signed-off-by: Samuel Holland <samuel@sholland.org>