drivers: IRQ polling does not need indirection
Like the counter, there is really only one IRQ chip implementation that
we care about, and there's no need to select one at runtime. So we can
avoid quite a bit of overhead by replacing the runtime indirection by
(if needed) compile-time selection of the correct code. Again, like the
counter, there is currently only one implementation, so it's not
suffixed.

Signed-off-by: Samuel Holland <samuel@sholland.org>