scpi: Only run the service while the system is running
This allows the msgbox clock to be turned off in suspend, once it has a
refcount of zero. It also makes the inactive polling loop a bit tighter.

This works because all of the SCPI commands that cause an
active->inactive state transition do not require a reply. (However, even
if they did, it would still get sent out; we just wouldn't see tx_full
go false until we started polling again.)

Signed-off-by: Samuel Holland <samuel@sholland.org>