drivers: clock: Move clock trees to device-specific files
This separates the clock controller implementation (which is shared
between all SoCs) from the description of the clock tree (which depends
on the SoC in use). This allows more flexibility to accurate represent
the differences between SoCs, and removes a layer of indirection between
clocks and their gate/reset bits.

Signed-off-by: Samuel Holland <samuel@sholland.org>