drivers: clock: Move clock lists out of platforms
List the known clocks per device instead of per platform. For clock
devices that are shared between platforms (e.g. the sun8i R_CCU), this
avoids duplicating the definition in each platform.

Since consumers of these clocks need to be independent of which device
is in use, config symbols are used to redirect to the appropriate
per-device clock list.

Signed-off-by: Samuel Holland <samuel@sholland.org>