device: Add a device release hook
This allows drivers to perform cleanup actions when a device is no
longer in use.

Since probe and release functions may not be needed for all drivers,
provide no-op implementations that can be reused across drivers. This is
actually more efficient in terms of code size than adding an if
statement to the generic function.

Signed-off-by: Samuel Holland <samuel@sholland.org>