drivers: clock: Move parent get_rate to generic code
The driver is only responsible for reading the factors specific to the
clock in question. Generic code handles traversing the clock tree. This
requires handling mux-specific dividers in a hook function, since the
driver's get_rate hook shouldn't be looking at the parent.

As a side effect, this removes the vdiv member from struct clock_handle,
and the FIXED_CLOCK macro (which was pretending to be driver-agnostic).

Signed-off-by: Samuel Holland <samuel@sholland.org>