common: Prohibit statically-initialized mutable data
When the firmare resets, it wipes .bss, but there is no way to reset
.data back to a pristine state. Therefore, we prohibit having a .data
section at all. All data must be constant (and placed in .rodata),
or it must be statically zeroed and initialized at runtime (and thus
allocated in .bss).

The firmware currently fulfills this requirement, but we can use a
linker script assertion to ensure this quality is maintained.

Signed-off-by: Samuel Holland <samuel@sholland.org>