drivers: clock: Return rate directly from clock_get_rate
Most of the time, getting a clock's frequency is not going to fail:
either because it is fixed (or a fixed divider) and cannot fail by
construction, or because it is MMIO and we wouldn't know if it failed.

In the rare case where reading a clock rate/divider fails, there is a
logical return value: 0. For everywhere else, pass rate by value instead
of by reference and save some bytes/boilerplate.

Signed-off-by: Samuel Holland <samuel@sholland.org>