lib: Refactor bitfield helpers to not use bitfield_t
Attempting to pack both the bitfield location and size in a single byte
was the cause of much of the code bloat in the clock drivers. This
change rewrites the functions to be more compiler-optimization-friendly.
Now, when the bitfield size/location is known at compile time, the
compiler can generate much better code. This will be the common case as
the CCU driver transitions to per-clock rate and parent hooks.

Signed-off-by: Samuel Holland <samuel@sholland.org>