device: Fix unbalanced reference counting
The refcount was not getting incremented for the second and following
calls to device_get().

Consumers of a device should only look at the return value of
device_get(). They should not be examining any device structs via
device_is_running() without a reference (and if they have a reference,
they know the device is running).

Signed-off-by: Samuel Holland <samuel@sholland.org>