tools: Remove unnecessary function attribute
When I rewrote the function attribute macros, I didn't check the tools.
The test tool still compiles because GCC treats __const as an alias for
const (the type qualifier, not the function attribute). While I could
fix the function attribute, the microoptimization (which is unlikely to
even make a difference at all) is unnecessary, so remove it.

Signed-off-by: Samuel Holland <samuel@sholland.org>