css: a64: Clean up register definitons
When this code was originally written, it was designed to be
future-proof to some extent, supporting an arbitrary number of clusters
and cores per cluster. The future is now here, with the H6, and it is
obvious that it is impossible to have multiple clusters with the current
register layout. In several cases, there is a register immediately
following a "cluster 0" register.

So let's drop the cluster parameter on all of the register macros. This
saves quite a bit of code space, since almost all register addresses are
now constant. There is still an assertion for the cluster number just in
case an invalid cluster is given. And the new behavior in that situation
is theoretically safer anyway -- the wrong core will be killed, but the
code will no longer clobber arbitrary adjacent registers.

Similarly, make the saved RVBA a scalar instead of an array. The CPU can
probably optimize out the indexing, but there's no real point to it.

The H6 SoC also moves both of the PRCM registers into R_CPUCFG and adds
a new MMIO area called CPU_SUBSYS_CTRL for CSS control outside the
cluster level. So to keep the register names as consistent as possible
inside the functions, move the device base address to the definition,
and remove the device name from the register name.

Finally, remove registers that are unrelated to the CSS (ARISC_RESET);
or that are only used by the BROM, which we bypass completely (SS_FLAG,
CPU_ENTRY, SS_ENTRY, HP_FLAG).

This commit should have no functional change.

Signed-off-by: Samuel Holland <samuel@sholland.org>