css: h6: Add basic CSS control for the H6 SoC
Other than rearranging register addresses, the core on/off sequence is
the same for H6 as it is for A64/H5. Core 0 on H6 does have its own
power switch, so it does not need to be skipped during CPU hotplug.

For cluster control, some resets have moved. H_RESET is no longer
present in CLUSTER_RESET_CTRL_REG, and some new resets have appeared.
Writing CPU_SYS_RESET_REG hangs, possibly because H_RESET is not
asserted. More investigation is needed.

Since it's not possible to turn off VDD-CPUX yet, this cannot be
considered a full suspend implementation, but it is plenty for Linux to
bring up the secondary cores, and successfully perform a deep ("mem")
suspend/resume cycle.

Signed-off-by: Samuel Holland <samuel@sholland.org>