system_power: Be safer about non-runtime resets
If we assume the system is off, any user interaction (IRQ) will cause a
full system reset. This is safer than attempting to resume the system,
especially if it is actually OFF, and additional clocks/regulators were
turned off during SHUTDOWN. In that case, the resume process may hang.
Or because external power domains (i.e. the SD card) are off, it may
cause a reset but be unable to boot.

Signed-off-by: Samuel Holland <samuel@sholland.org>